(Gabriel Costa 05/07/2023) IncidenciaDbMapping test
