+++ Tests/UnitTests/Web.Api.BocaSuja/Context/BocaSujaDbContextTest.cs
@@ -59,42 +59,4 @@ public class BocaSujaDbContextTest
         });
     }
 
-    [Test, Category("WebApi - Context - BocaSujaDbContext - InstanciaDbMapping")]
-    [Description("Verifica a instancia do.")]
-    public void InstantiateDbContext(string validRecurso, string validTipo, int validGravidade)
-    {
-        using var dbContext = new BocaSujaDbContext(_options);
-        var entidadeOfensoraGuid = Guid.NewGuid();
-
-        var incidencia = new Incidencia(
-            entidadeOfensoraGuid,
-            validRecurso,
-            validTipo,
-            validGravidade
-        );
-
-        dbContext.Incidencias.Add(incidencia);
-        dbContext.SaveChanges();
-
-        var incidenciaPersistida = dbContext.Incidencias.Find(incidencia.Id);
-
-        Assert.Multiple(() =>
-        {
-            Assert.That(actual: dbContext.Incidencias.Count(), Is.AtLeast(1));
-            Assert.That(actual: incidenciaPersistida, Is.InstanceOf(typeof(Incidencia)));
-            Assert.That(actual: incidenciaPersistida, Is.Not.Null);
-            Assert.That(actual: incidenciaPersistida?.Id, Is.InstanceOf(typeof(Guid)));
-            Assert.That(
-                actual: incidenciaPersistida?.EntidadeOfensora,
-                Is.EqualTo(incidencia.EntidadeOfensora)
-            );
-            Assert.That(actual: incidenciaPersistida?.Recurso, Is.EqualTo(incidencia.Recurso));
-            Assert.That(actual: incidenciaPersistida?.Tipo, Is.EqualTo(incidencia.Tipo));
-            Assert.That(actual: incidenciaPersistida?.Gravidade, Is.EqualTo(incidencia.Gravidade));
-            Assert.That(
-                actual: incidenciaPersistida?.DataHoraCriacao,
-                Is.EqualTo(incidencia.DataHoraCriacao)
-            );
-        });
-    }
 }
