--- Tests/UnitTests/Web.Api.BocaSuja/Context/DbMapping/IncidenciaDbMapping.cs
@@ -1,19 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace Tests.UnitTests.Web.Api.BocaSuja.Context.DbMapping;
-
-[TestFixture]
-public class IncidenciaDbMapping
-{
-
-    private ModelBuilder _modelBuilder;
-    private IncidenciaDbMapping _incidenciaDbMapping;
-
-    [SetUp]
-    public void Setup()
-    {
-        _modelBuilder = new ModelBuilder();
-        _incidenciaDbMapping = new IncidenciaDbMapping();
-    }
-
-}
